Renamed.
I think the occurrence module name is more descriptive for what we're
trying to achieve here.
